Create README
Use an 8V power supply to reduce voltage and supply power to the Raspberry Pi and relay module, a 12V power supply to supply power to the latch, and the 24, 25, 27, and 28 IO ports of the Raspberry Pi generate 3.3V voltage to input signals to the four relay modules. The relay shares ground with the Raspberry Pi. When facial recognition is successful, the relay module operates and the 12V power supply begins to supply power to the latch.